requirement.txt
twisted
protobuf
txmongo
motor
uvloop
aioredis
pprint
arrow
pillow